# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Atlas Payments API
  version: 0.2.0
  description: Core payment processing — intents, captures, refunds, balance, and payment methods
  contact:
    name: AtlasPayments Team
    url: https://github.com/shivemind/atlaspayments
servers:
  - url: http://localhost:3000
    description: Local development server. Override via Postman {{baseUrl}} environment variable.
...
paths:
...
  /api/v1/setup_intents:
    get:
      summary: List setup intents
      description: Returns a paginated list of setup intents for the authenticated merchant.
      x-platform:
        kind: resource
        entity: SetupIntent
        affectsLedger: false
      operationId: listSetupIntents
      tags:
        - SetupIntents
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
        - in: query
          name: customer_id
          description: Filter by customer ID.
          schema:
            type: string
        - in: query
          name: status
          description: Filter by setup intent status.
          schema:
            type: string
            enum:
              - created
              - processing
              - succeeded
              - failed
              - canceled
      responses:
        "200":
          description: Paginated list of setup intents.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
              examples:
                page:
                  value:
                    data:
                      - id: cmf1si123
                        merchant_id: cmf1merchant123
                        customer_id: cmf1customer123
                        status: created
                        usage: off_session
                    pagination:
                      page: 1
                      pageSize: 10
                      total: 1
                      hasMore: false
        "400":
          description: Invalid query parameters.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create setup intent
      description: >
        Creates a setup intent in `created` status for attaching a payment method to a customer. Emits
        `setup_intent.created`.
      x-platform:
        kind: transition
        entity: SetupIntent
        stateChanges:
          - created
        emitsEvents:
          - setup_intent.created
        affectsLedger: false
        workflowRefs:
          - attach_payment_method
          - subscription_lifecycle
      operationId: createSetupIntent
      tags:
        - SetupIntents
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/IdempotencyKey"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - customer_id
              properties:
                customer_id:
                  type: string
                  description: ID of the customer for this setup intent.
                usage:
                  type: string
                  enum:
                    - off_session
                    - on_session
                  default: off_session
                  description: How the payment method will be used.
                metadata:
                  type: object
                  additionalProperties: true
            examples:
              basic:
                value:
                  customer_id: "{{customerId}}"
                  usage: off_session
      responses:
        "201":
          description: Setup intent created in `created` status.
          content:
            application/json:
              schema:
                type: object
                required:
                  - setup_intent
                properties:
                  setup_intent:
                    type: object
                    required:
                      - id
                      - merchant_id
                      - customer_id
                      - status
                      - usage
                      - created_at
                      - updated_at
                    properties:
                      id:
                        type: string
                      merchant_id:
                        type: string
                      customer_id:
                        type: string
                      payment_method_id:
                        type: string
                        nullable: true
                      status:
                        type: string
                        enum:
                          - created
                          - processing
                          - succeeded
                          - failed
                          - canceled
                      usage:
                        type: string
                      cancellation_reason:
                        type: string
                        nullable: true
                      metadata:
                        type: object
                        nullable: true
                      created_at:
                        type: string
                        format: date-time
                      updated_at:
                        type: string
                        format: date-time
              examples:
                created:
                  value:
                    setup_intent:
                      id: cmf1si123
                      merchant_id: cmf1merchant123
                      customer_id: "{{customerId}}"
                      payment_method_id: null
                      status: created
                      usage: off_session
                      cancellation_reason: null
                      metadata: null
                      created_at: 2026-03-01T10:00:00.000Z
                      updated_at: 2026-03-01T10:00:00.000Z
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Customer not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Idempotency key conflict.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/v1/refunds:
    get:
      summary: List refunds
      description: Returns a paginated list of refunds for the authenticated merchant.
      x-platform:
        kind: resource
        entity: Refund
        affectsLedger: false
      operationId: listRefunds
      tags:
        - Refunds
      security:
        - MerchantKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/PageSizeParam"
      responses:
        "200":
          description: Paginated list of merchant refunds.
          content:
            application/json:
              schema:
                type: object
                required:
                  - data
                  - pagination
                properties:
                  data:
                    type: array
                    items:
                      type: object
                  pagination:
                    $ref: "#/components/schemas/PaginationMeta"
              examples:
                page:
                  value:
                    data:
                      - id: cmf1refund123
                        amount: 2500
                        status: succeeded
                    pagination:
                      page: 1
                      pageSize: 10
                      total: 1
                      hasMore: false
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      summary: Create refund
      description: >
        Creates a refund for a captured payment intent (partial supported). Posts ledger reversal entries and emits
        `refund.created` + `refund.succeeded`.
      x-platform:
        kind: transition
        entity: Refund
        stateChanges:
          - created -> succeeded
        emitsEvents:
          - refund.created
          - refund.succeeded
        affectsLedger: true
        workflowRefs:
          - payment_and_refund
      operationId: createRefund
      tags:
        - Refunds
      security:
        - MerchantKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - payment_intent_id
                - amount
              properties:
                payment_intent_id:
                  type: string
                amount:
                  type: integer
                  minimum: 1
                reason:
                  type: string
            examples:
              basic:
                value:
                  payment_intent_id: "{{paymentIntentId}}"
                  amount: 2500
                  reason: requested_by_customer
      responses:
        "201":
          description: Refund created and immediately succeeded.
          content:
            application/json:
              schema:
                type: object
                properties:
                  refund:
                    type: object
                    properties:
                      id:
                        type: string
                      merchant_id:
                        type: string
                      payment_intent_id:
                        type: string
                      amount:
                        type: integer
                      status:
                        type: string
                      reason:
                        type: string
                        nullable: true
                      created_at:
                        type: string
                        format: date-time
              examples:
                refunded:
                  value:
                    refund:
                      id: cmf1refund123
                      merchant_id: cmf1merchant123
                      payment_intent_id: "{{paymentIntentId}}"
                      amount: 2500
                      status: succeeded
                      reason: requested_by_customer
                      created_at: 2026-03-01T10:00:00.000Z
        "400":
          description: Invalid request payload.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid API key.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Payment intent not found.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Not captured or amount exceeds refundable balance.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal server error.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    MerchantKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for merchant-scoped operations. Pass as `Authorization: Bearer {{merchantApiKey}}`.
    PlatformAdminKeyAuth:
      type: http
      scheme: bearer
      bearerFormat: API Key
      description: |
        Bearer API key for platform admin operations. Pass as `Authorization: Bearer {{platformAdminKey}}`.
  schemas:
    ErrorResponse:
      description: Standard error response envelope returned for 4xx/5xx responses.
      type: object
      required:
        - error
      properties:
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              description: Machine-readable error code.
            message:
              type: string
              description: Human-readable error description.
    PaginationMeta:
      description: Pagination metadata included in list responses.
      type: object
      required:
        - page
        - pageSize
        - total
        - hasMore
      properties:
        page:
          type: integer
          description: Current page number (1-indexed).
        pageSize:
          type: integer
          description: Number of results per page.
        total:
          type: integer
          description: Total number of matching records.
        hasMore:
          type: boolean
          description: Whether additional pages exist.
...
  parameters:
    IdempotencyKey:
      in: header
      name: Idempotency-Key
      required: true
      description: |
        Client-generated unique key for safe request retries. Identical key + payload replays the original response.
      schema:
        type: string
        example: idem-key-a1b2c3d4
    PageParam:
      in: query
      name: page
      description: Page number (1-indexed).
      schema:
        type: integer
        minimum: 1
        default: 1
    PageSizeParam:
      in: query
      name: pageSize
      description: Number of results per page (max 100).
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
...